+++ lib/spec/shared_examples/simple_crud_for_destroy.rb
@@ -4,7 +4,8 @@ shared_examples 'simple crud for destroy' do
   describe 'DELETE #destroy' do
     context 'without authenticated user' do
       subject!(:req) do
-        delete :destroy, params: record_param(:destroy, nil, not_found: true), format: request_format(:destroy)
+        delete :destroy, params: with_route_params(record_param(:destroy, nil, not_found: true)),
+                         format: request_format(:destroy)
       end
 
       include_examples 'unauthorized when not logged in' if check_authenticate(:destroy)
@@ -16,7 +17,7 @@ shared_examples 'simple crud for destroy' do
 
         before do
           make_policies_fail(:destroy)
-          delete :destroy, params: record_param(:destroy, model), format: request_format(:destroy)
+          delete :destroy, params: with_route_params(record_param(:destroy, model)), format: request_format(:destroy)
         end
 
         it 'fails with forbidden' do
@@ -30,7 +31,7 @@ shared_examples 'simple crud for destroy' do
 
       before do
         model
-        delete :destroy, params: record_param(:destroy, model), format: request_format(:destroy)
+        delete :destroy, params: with_route_params(record_param(:destroy, model)), format: request_format(:destroy)
       end
 
       if check_html(:destroy)
@@ -50,7 +51,8 @@ shared_examples 'simple crud for destroy' do
 
       before do
         model
-        delete :destroy, params: record_param(:destroy, nil, not_found: true), format: request_format(:destroy)
+        delete :destroy, params: with_route_params(record_param(:destroy, nil, not_found: true)),
+                         format: request_format(:destroy)
       end
 
       it 'responds with not found status' do
@@ -65,7 +67,7 @@ shared_examples 'simple crud for destroy' do
         before do
           model
           make_policies_fail(:destroy)
-          delete :destroy, params: record_param(:destroy, model), format: request_format(:destroy)
+          delete :destroy, params: with_route_params(record_param(:destroy, model)), format: request_format(:destroy)
         end
 
         it 'responds with forbidden status' do