+++ lib/spec/shared_examples/simple_crud_for_new_with_block.rb
@@ -2,7 +2,7 @@
 
 shared_examples 'simple crud for new with block' do
   describe 'GET #new with a render block' do
-    before { get :new, format: :html }
+    before { get :new, params: with_route_params({}), format: :html }
 
     include_examples 'simple crud renders template', :new
   end