+++ lib/spec/shared_examples/simple_crud_for_new_with_build.rb
@@ -4,7 +4,7 @@ shared_examples 'simple crud for new with build' do
   describe 'GET #new with a build hook' do
     include_context 'with authenticated user'
 
-    before { get :new }
+    before { get :new, params: with_route_params({}) }
 
     it 'builds the record owned by the current user' do
       expect(response_body[owner_foreign_key.to_s]).to eq(current_user.id)