+++ lib/spec/shared_examples/simple_crud_for_show_with_block.rb
@@ -4,7 +4,7 @@ shared_examples 'simple crud for show with block' do
   describe 'GET #show with a render block' do
     let!(:record) { model }
 
-    before { get :show, params: { id: record.id }, format: :html }
+    before { get :show, params: with_route_params({ id: record.id }), format: :html }
 
     include_examples 'simple crud renders template', :show
   end