+++ lib/spec/shared_examples/simple_crud_without_authenticated_user.rb
@@ -2,7 +2,7 @@
 
 shared_examples 'simple crud without authenticated user' do |action|
   context 'without authenticated user' do
-    subject!(:req) { get action }
+    subject!(:req) { get action, params: with_route_params({}) }
 
     include_examples 'unauthorized when not logged in' if check_authenticate(action)
   end